+++ R/data.R
@@ -19,6 +19,7 @@
 #'  
 #' @docType data
 #' @name immapex_AA.data
+#' @return List of 10 amino acid properties for 20 amino acids
 #' 
 NULL
 
@@ -52,6 +53,7 @@ NULL
 #' 
 #' @docType data
 #' @name immapex_example.data
+#' @return List of 3 example data sets for 10x, AIRR and Adaptive contigs.
 #' 
 NULL
 
@@ -77,5 +79,6 @@ NULL
 #' 
 #' @docType data
 #' @name immapex_blosum.pam.matrices
+#' @return List of 10 substitution matrices
 #' 
 NULL