+++ R/formatGenes.R
@@ -16,6 +16,7 @@
 #' retain all information (\strong{FALSE})
 #' 
 #' @examples
+#' data(immapex_example.data)
 #' formatGenes(immapex_example.data[["TenX"]],
 #'             region = "v",
 #'             technology = "TenX")
@@ -63,7 +64,7 @@ formatGenes <- function(input.data,
     }
   }
  
-  
+  data("immapex_gene.list", package = "immApex", envir = environment())
   gene.list <- immapex_gene.list[tolower(region)]
   lapply(gene.list, function(x) {
     lapply(x, function(y) {