+++ R/inferCDR.R
@@ -7,6 +7,7 @@
 #' 
 #' @examples
 #' # Getting the Sequence Reference
+#' data(immapex_example.data)
 #' TRBV_aa <- getIMGT(species = "human",
 #'                    chain = "TRB",
 #'                    frame = "inframe",